feat: all tests pass with mock
